fix(Interaction): correctly set attach point when switching sdk
There was an issue where if SteamVR was not the first SDK used and
the SDK was later switched to SteamVR, the controller attach point
on InteractGrab would not get updated due to the slight lag that
SteamVR has when initialising the controllers.

This slight lag meant that the controller model was not populated
for a fraction of a second when the InteractGrab script attempted
to populate the controller attach point, meaning it would just
stay as the previous attach point which was incorrect.

This fix uses a coroutine to check a limited number of times for
the existence of a valid attach path over a period of time.

It also does this check when the controller index changes in case
the index change also requires an update to the attach point.